+++ core/src/main/scala/org/apache/spark/sql/streaming/eventhubs/EventHubsSource.scala
@@ -58,23 +58,21 @@ private[spark] class EventHubsSource private[eventhubs] (
   require(eventHubsNamespace != null, "eventhubs.namespace is not defined")
   require(eventHubsName != null, "eventhubs.name is not defined")
 
-  private var _eventHubsClient: Client = _
-
-  private var _eventHubsReceiver: (Map[String, String]) => Client = _
-
+  private var client: Client = _
   private[eventhubs] def eventHubClient = {
-    if (_eventHubsClient == null) {
-      _eventHubsClient = clientFactory(eventHubsParams)
-    }
-    _eventHubsClient
+    if (client == null) client = clientFactory(eventHubsParams)
+    client
   }
+  private[spark] def eventHubClient_=(eventHubClient: Client) =
+    client = eventHubClient
 
+  private var receiverFactory: (Map[String, String]) => Client = _
   private[eventhubs] def eventHubsReceiver = {
-    if (_eventHubsReceiver == null) {
-      _eventHubsReceiver = clientFactory
-    }
-    _eventHubsReceiver
+    if (receiverFactory == null) clientFactory
+    receiverFactory
   }
+  private[spark] def eventHubsReceiver_=(factory: (Map[String, String]) => Client) =
+    receiverFactory = factory
 
   private val ehNameAndPartitions = {
     val partitionCount = eventHubsParams("eventhubs.partition.count").toInt
@@ -95,17 +93,6 @@ private[spark] class EventHubsSource private[eventhubs] (
     eventHubsParams("eventhubs.progressTrackingDir"),
     sqlContext.sparkContext.appName,
     sqlContext.sparkContext.hadoopConfiguration)
-
-  private[spark] def setEventHubClient(eventHubClient: Client): EventHubsSource = {
-    _eventHubsClient = eventHubClient
-    this
-  }
-
-  private[spark] def setEventHubsReceiver(
-      eventhubReceiverCreator: (Map[String, String]) => Client): EventHubsSource = {
-    _eventHubsReceiver = eventhubReceiverCreator
-    this
-  }
 
   // the flag to avoid committing in the first batch
   private[spark] var firstBatch = true
