+++ core/src/main/scala/org/apache/spark/streaming/eventhubs/EventHubDirectDStream.scala
@@ -81,25 +81,22 @@ private[eventhubs] class EventHubDirectDStream private[eventhubs] (
    */
   }
 
-  @transient private var _eventHubClients: mutable.HashMap[String, Client] = _
-
   private def progressTracker =
     DirectDStreamProgressTracker.getInstance.asInstanceOf[DirectDStreamProgressTracker]
 
-  private[eventhubs] def setEventHubClient(
-      eventHubClient: mutable.HashMap[String, Client]): EventHubDirectDStream = {
-    _eventHubClients = eventHubClient
-    this
-  }
+  @transient private var clients: mutable.HashMap[String, Client] = _
 
-  private[eventhubs] def eventHubClient = {
-    if (_eventHubClients == null) {
-      _eventHubClients = new mutable.HashMap[String, Client].empty
+  private[eventhubs] def eventHubClients_=(clients: mutable.HashMap[String, Client]) =
+    this.clients = clients
+
+  private[eventhubs] def eventHubClients = {
+    if (this.clients == null) {
+      this.clients = new mutable.HashMap[String, Client].empty
       for (name <- eventhubsParams.keys)
-        _eventHubClients += name -> clientFactory(eventhubsParams(name))
+        this.clients += name -> clientFactory(eventhubsParams(name))
     }
 
-    _eventHubClients
+    this.clients
   }
 
   private[eventhubs] var currentOffsetsAndSeqNums = OffsetRecord(-1L, {
@@ -149,7 +146,7 @@ private[eventhubs] class EventHubDirectDStream private[eventhubs] (
       val startSeqs = new mutable.HashMap[EventHubNameAndPartition, Long].empty
       for (nameAndPartition <- connectedInstances) {
         val name = nameAndPartition.eventHubName
-        val seqNo = eventHubClient(name).beginSeqNo(nameAndPartition)
+        val seqNo = eventHubClients(name).beginSeqNo(nameAndPartition)
         require(seqNo.isDefined, s"Failed to get starting sequence number for $nameAndPartition")
 
         startSeqs += nameAndPartition -> seqNo.get
@@ -239,7 +236,7 @@ private[eventhubs] class EventHubDirectDStream private[eventhubs] (
     val filteringOffsetAndType = {
       if (shouldCareEnqueueTimeOrOffset) {
         // first check if the parameters are valid
-        RateControlUtils.validateFilteringParams(eventHubClient.toMap,
+        RateControlUtils.validateFilteringParams(eventHubClients.toMap,
                                                  eventhubsParams,
                                                  connectedInstances)
         RateControlUtils.composeFromOffsetWithFilteringParams(eventhubsParams,
@@ -313,7 +310,7 @@ private[eventhubs] class EventHubDirectDStream private[eventhubs] (
 
   private[spark] def composeHighestOffset(validTime: Time, retryIfFail: Boolean) = {
     RateControlUtils.fetchLatestOffset(
-      eventHubClient.toMap,
+      eventHubClients.toMap,
       if (fetchedHighestOffsetsAndSeqNums == null) {
         currentOffsetsAndSeqNums.offsets
       } else {
