+++ core/src/test/scala/org/apache/spark/eventhubscommon/utils/TestClients.scala
@@ -1,0 +1,144 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.spark.eventhubscommon.utils
+
+import com.microsoft.azure.eventhubs.{ EventData, EventHubClient }
+import org.apache.spark.eventhubscommon.EventHubNameAndPartition
+import org.apache.spark.eventhubscommon.client.{ Client, EventHubsOffsetTypes }
+import org.apache.spark.eventhubscommon.client.EventHubsOffsetTypes.EventHubsOffsetType
+import org.apache.spark.streaming.StreamingContext
+
+// TODO: consolidate these. We should only need two.
+class SimulatedEventHubsRestClient(eventHubs: SimulatedEventHubs)
+    extends Client
+    with TestClientSugar {
+
+  override def lastSeqAndOffset(
+      nameAndPartition: EventHubNameAndPartition): Option[(Long, Long)] = {
+    val x = eventHubs.messageStore(nameAndPartition).length.toLong - 1
+    Some((x, x))
+  }
+
+  override def lastEnqueuedTime(nameAndPartition: EventHubNameAndPartition): Option[Long] = {
+    Some(
+      eventHubs
+        .messageStore(nameAndPartition)
+        .last
+        .getSystemProperties
+        .getEnqueuedTime
+        .toEpochMilli)
+  }
+}
+
+class TestEventHubsClient(ehParams: Map[String, String],
+                          eventHubs: SimulatedEventHubs,
+                          latestRecords: Map[EventHubNameAndPartition, (Long, Long, Long)])
+    extends Client
+    with TestClientSugar {
+  override def receive(expectedEventNum: Int): Iterable[EventData] = {
+    val eventHubName = ehParams("eventhubs.name")
+    if (offsetType != EventHubsOffsetTypes.EnqueueTime) {
+      eventHubs.search(EventHubNameAndPartition(eventHubName, partitionId),
+                       currentOffset.toInt,
+                       expectedEventNum)
+    } else {
+      eventHubs.searchWithTime(EventHubNameAndPartition(eventHubName, partitionId),
+                               ehParams("eventhubs.filter.enqueuetime").toLong,
+                               expectedEventNum)
+    }
+  }
+
+  override def lastSeqAndOffset(
+      nameAndPartition: EventHubNameAndPartition): Option[(Long, Long)] = {
+    val (offset, seq, _) = latestRecords(nameAndPartition)
+    Some(offset, seq)
+  }
+
+  override def lastEnqueuedTime(nameAndPartition: EventHubNameAndPartition): Option[Long] =
+    Some(latestRecords(nameAndPartition)._3)
+}
+
+class FluctuatedEventHubClient(ehParams: Map[String, String],
+                               eventHubs: SimulatedEventHubs,
+                               ssc: StreamingContext,
+                               messagesBeforeEmpty: Long,
+                               numBatchesBeforeNewData: Int,
+                               latestRecords: Map[EventHubNameAndPartition, (Long, Long)])
+    extends Client
+    with TestClientSugar {
+  private var callIndex = -1
+
+  override def receive(expectedEventNum: Int): Iterable[EventData] = {
+    val eventHubName = ehParams("eventhubs.name")
+    if (offsetType != EventHubsOffsetTypes.EnqueueTime) {
+      eventHubs.search(EventHubNameAndPartition(eventHubName, partitionId),
+                       currentOffset.toInt,
+                       expectedEventNum)
+    } else {
+      eventHubs.searchWithTime(EventHubNameAndPartition(eventHubName, partitionId),
+                               ehParams("eventhubs.filter.enqueuetime").toLong,
+                               expectedEventNum)
+    }
+  }
+
+  override def lastSeqAndOffset(
+      nameAndPartition: EventHubNameAndPartition): Option[(Long, Long)] = {
+    callIndex += 1
+    if (callIndex < numBatchesBeforeNewData) {
+      Some(messagesBeforeEmpty - 1, messagesBeforeEmpty - 1)
+    } else {
+      Some(latestRecords(nameAndPartition))
+    }
+  }
+
+  override def lastEnqueuedTime(nameAndPartition: EventHubNameAndPartition): Option[Long] = {
+    Some(Long.MaxValue)
+  }
+}
+
+/**
+ * TestClientSugar implements Client so all methods and variables are NOPs. This will
+ * reduce repetitive code when the Client is mocked in testing.
+ */
+sealed trait TestClientSugar extends Client {
+  override private[spark] var client: EventHubClient = _
+  protected var partitionId: Int = _
+  protected var offsetType: EventHubsOffsetType = _
+  protected var currentOffset: String = _
+
+  override def initReceiver(partitionId: String,
+                            offsetType: EventHubsOffsetType,
+                            currentOffset: String): Unit = {
+    this.partitionId = partitionId.toInt
+    this.offsetType = offsetType
+    this.currentOffset = currentOffset
+  }
+
+  override def close(): Unit = {}
+
+  override def lastSeqAndOffset(
+      eventHubNameAndPartition: EventHubNameAndPartition): Option[(Long, Long)] = Option.empty
+
+  override def lastEnqueuedTime(eventHubNameAndPartition: EventHubNameAndPartition): Option[Long] =
+    Option.empty
+
+  override def receive(expectedEvents: Int): Iterable[EventData] = Iterable[EventData]()
+
+  override def beginSeqNo(eventHubNameAndPartition: EventHubNameAndPartition): Option[Long] =
+    Some(-1L)
+}
