+++ core/src/test/scala/org/apache/spark/sql/streaming/eventhubs/EventHubsStreamTest.scala
@@ -38,7 +38,6 @@ import org.scalatest.time.SpanSugar._
 
 import org.apache.spark.DebugFilesystem
 import org.apache.spark.eventhubscommon.EventHubsConnector
-import org.apache.spark.eventhubscommon.client.EventHubsOffsetTypes.EventHubsOffsetType
 import org.apache.spark.eventhubscommon.progress.ProgressTrackerBase
 import org.apache.spark.eventhubscommon.utils._
 import org.apache.spark.sql.{ Dataset, Encoder, QueryTest, Row }
@@ -511,11 +510,9 @@ trait EventHubsStreamTest
 
             val eventHubsSource = searchCurrentSource()
             val eventHubs = EventHubsTestUtilities.getOrSimulateEventHubs(null)
-            eventHubsSource.setEventHubClient(new SimulatedEventHubsRestClient(eventHubs))
-            eventHubsSource.setEventHubsReceiver(
-              (eventHubsParameters: Map[String, String]) =>
-                new TestEventHubsClient(eventHubsParameters, eventHubs, null)
-            )
+            eventHubsSource.eventHubClient = new SimulatedEventHubsRestClient(eventHubs)
+            eventHubsSource.eventHubsReceiver =
+              (ehParams: Map[String, String]) => new TestEventHubsClient(ehParams, eventHubs, null)
             currentStream.start()
 
           case AdvanceManualClock(timeToAdd) =>
